api/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
django-split-settings==1.3.1
djangorestframework==3.15.1
environs==11.0.0
marshmallow==3.21.1
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
